refactored generateLegendItems.tsx
